% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.adaboost.R
\name{print.real_adaboost}
\alias{print.real_adaboost}
\title{Print real adaboost model summary}
\usage{
\method{print}{real_adaboost}(x, ...)
}
\arguments{
\item{x}{object of class real_adaboost}

\item{...}{arguments passed to print.default}
}
\value{
None
}
\description{
S3 method to print a real_adaboost object
}
\details{
Displays basic information on the model,
such as function call, dependent variable
and the number of trees
}
\examples{
fakedata <- data.frame( X=c(rnorm(100,0,1),rnorm(100,1,1)), Y=c(rep(0,100),rep(1,100) ) )
fakedata$Y <- factor(fakedata$Y)
test_real_adaboost<- real_adaboost(Y~X, fakedata, 10)
print(test_real_adaboost)
}
\seealso{
\code{\link{print.adaboost}}
}

